[hardfork] Create a devnet only pre-portland hard fork (#223)
# Description

A newer node in the `devnet` could not sync to the latest block, which
verify block failed on height `1062992`.

The hard fork was brought up to the network a little rushed, which is
named `portland`, too.

This PR fixes it. We rename it to the `pre-portland` and it is only
available in some test networks.

# Changes include

- [x] Bugfix (non-breaking change that solves an issue)

# Checklist

- [x] I have assigned this PR to myself
- [x] I have added at least 1 reviewer
- [x] I have added the relevant labels

## Testing

- [x] I have tested this code with the official test suite
- [x] I have tested this code manually

### Manual tests

* Set up a new `devnet` node, and join in the network.
* Watch if it can catch up to the latest block height.

It works in this branch, but fail in the base branch.