[ibft] Fix ddos mark count increase insanely with only one transaction (#257)
# Description

The PR fixes a bug imported by #254.
A marked "DDOS" transaction would block the whole block building work,
and mark insanely until it was dropped.
The mark count is not right, due to not skip by (`Pop()` in fact).

# Changes include

- [x] Bugfix (non-breaking change that solves an issue)

## Testing

- [x] I have tested this code with the official test suite
- [x] I have tested this code manually

### Manual tests

1. Set up a 4-validator network, all enable `ddosProtection`.
2. Deploy CoinTool contact, and call it with some scripts.
3. Send 6 transactions all together to one validator.
4. Check whether it is able to send more contract transactions to the
validator, after 1 minutes.

In the PR branch, it is OK to go on with step 4. In contrast, the base
branch would failed on it.